+++ test/unit/gen_class/timeformat_init.py
@@ -16,7 +16,6 @@
 # Standard
 import sys
 import os
-import datetime
 import unittest
 
 # Local
@@ -35,8 +34,6 @@ class UnitTest(unittest.TestCase):
 
     Methods:
         setUp
-        test_rdtg
-        test_msecs
         test_delimit
         test_micro
         test_thacks
@@ -60,31 +57,10 @@ class UnitTest(unittest.TestCase):
         self.thacks = dict()
         self.tformats = {
             "ymd": {"format": "%Y%m%d", "del": "", "micro": False},
-            "dmy": {"format": "%d%m%Y", "del": "", "micro": False}}
-
-    def test_rdtg(self):
-
-        """Function:  test_rdtg
-
-        Description:  Test the rdtg attribute.
-
-        Arguments:
-
-        """
-
-        self.assertTrue(isinstance(self.tform.rdtg, datetime.datetime))
-
-    def test_msecs(self):
-
-        """Function:  test_msecs
-
-        Description:  Test the msecs attribute.
-
-        Arguments:
-
-        """
-
-        self.assertTrue(isinstance(self.tform.msecs, str))
+            "dmy": {"format": "%d%m%Y", "del": "", "micro": False},
+            "zulu": {
+                "format": "%Y-%m-%dT%H:%M:%SZ", "del": "", "micro": False},
+            "dtg": {"format": "%Y%m%d_%H%M%S", "del": "", "micro": False}}
 
     def test_delimit(self):