+++ CHANGELOG.md
@@ -4,7 +4,7 @@ All notable changes to this project will be documented in this file.
 The format is based on "Keep a Changelog".  This project adheres to Semantic Versioning.
 
 
-## [2.9.5] - 2023-01-17
+## [2.10.0] - 2023-01-17
 - Added new class called Daemon2.
 
 ### Fixed